+++ app/Modules/Payments/Interfaces/Resources/CashSessionResource.php
@@ -16,7 +16,7 @@ class CashSessionResource extends JsonResource
             'opening_amount' => (int) $this->opening_amount,  // ADR-011: integer CLP
             'closing_amount' => $this->closing_amount ? (int) $this->closing_amount : null,  // ADR-011: integer CLP
             'expected_amount' => $this->expected_amount ? (int) $this->expected_amount : null,  // ADR-011: integer CLP
-            'difference' => $this->difference ? (float) $this->difference : null,
+            'difference' => $this->difference ? (int) $this->difference : null,  // ADR-018
             'opening_notes' => $this->opening_notes,
             'closing_notes' => $this->closing_notes,
             'opened_at' => $this->opened_at?->toIso8601String(),
